upgrade solidity to 0.8.2

+++ contracts/interfaces/IUniswapV2Callee.sol
@@ -1,5 +1,12 @@
+// SPDX-License-Identifier: GPL-3.0-or-later
+
 pragma solidity >=0.5.0;
 
 interface IUniswapV2Callee {
-    function uniswapV2Call(address sender, uint amount0, uint amount1, bytes calldata data) external;
+    function uniswapV2Call(
+        address sender,
+        uint256 amount0,
+        uint256 amount1,
+        bytes calldata data
+    ) external;
 }
